Added pages and routing
