+++ app/Log-in/page.js
@@ -1,0 +1,9 @@
+import { Box, Typography } from "@mui/material";
+
+export default function Login() {
+  return (
+    <Box>
+      <Typography>Login</Typography>
+    </Box>
+  );
+}
